Separate test classes into separate files
