+++ app/src/androidTest/java/com/github/meeplemeet/ui/FriendsScreenTest.kt
@@ -699,63 +699,3 @@ class FriendsScreenTest : FirestoreTests() {
     }
   }
 }
-
-class FriendsScreenCrashTest : FirestoreTests() {
-
-  @get:Rule val compose = createComposeRule()
-
-  private lateinit var viewModel: FriendsScreenViewModel
-  private lateinit var currentUser: Account
-
-  @Before
-  fun setup() {
-    viewModel = FriendsScreenViewModel(accountRepository, handlesRepository)
-
-    runBlocking {
-      val suffix = System.currentTimeMillis()
-
-      currentUser =
-          accountRepository.createAccount(
-              userHandle = "crash_test_user_$suffix",
-              name = "Crash Test User",
-              email = "crash_$suffix@meeple.test",
-              photoUrl = null)
-
-      val friendToDelete =
-          accountRepository.createAccount(
-              userHandle = "to_be_deleted_$suffix",
-              name = "Deleted User",
-              email = "deleted_$suffix@meeple.test",
-              photoUrl = null)
-
-      accountRepository.sendFriendRequest(currentUser, friendToDelete.uid)
-      accountRepository.acceptFriendRequest(friendToDelete.uid, currentUser.uid)
-
-      // Now delete the friend account
-      accountRepository.deleteAccount(friendToDelete.uid)
-
-      currentUser = accountRepository.getAccount(currentUser.uid)
-    }
-  }
-
-  @OptIn(ExperimentalTestApi::class)
-  @Test
-  fun friendsScreen_missingFriend_doesNotCrash() {
-    compose.setContent {
-      AppTheme(themeMode = ThemeMode.DARK) {
-        FriendsScreen(
-            account = currentUser,
-            viewModel = viewModel,
-            onBack = {},
-        )
-      }
-    }
-
-    compose.waitUntilAtLeastOneExists(
-        hasTestTag(FriendsManagementTestTags.SCREEN_ROOT),
-        timeoutMillis = 5_000,
-    )
-
-    compose.onNodeWithTag(FriendsManagementTestTags.SCREEN_ROOT).assertIsDisplayed()
-  }
-}
